Removed MAX_ATTEMPTS and added get_total_time
Removed the limit on the maximum number of attempts a person can make per question.
Added method 'get_total_time' to Player. It gets the sum of times taken by player to solve questions correctly. This will be used as a tie-breaker in the leaderboard.
Added the view_ques permission.